requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
textblob==0.17.1
python-multipart==0.0.6
openai==1.12.0
python-dotenv==1.0.0
